comparing an integer with an unsigned integer FIX

+++ robot_patrol/src/patrol.cpp
@@ -25,8 +25,8 @@ private:
   {
     //Identify the largest distance ray, which is not inf
     float max_distance = 0;
-    int max_index = 0;
-    for (int i = 0; i < data_laser_->ranges.size(); i++)
+    size_t max_index = 0;
+    for (size_t i = 0; i < data_laser_->ranges.size(); i++)
     {
       if (data_laser_->ranges[i] > max_distance && data_laser_->ranges[i] < 10)
       {
